fix: split inbox storage helpers below size ceiling
Agend-Agent: archfix-codex-dev
Agend-Task: t-20260713222724486449-70517-167
Agend-Branch: fix/ci-handoff-episode-settlement
Agend-Issued-At: 2026-07-14T23:32:23.851412+00:00
